ai_pipeline/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
openai
langchain
langchain-core
langchain-community
langchain-openai
faiss-cpu
pypdf
cachetools
